Refactored code, updated dev tooling and fixed items not being consumed
