+++ .gitattributes
@@ -1,17 +1,75 @@
-# Auto detect text files and perform LF normalization
-* text=auto
+*            text=auto
 
-# Custom for Visual Studio
-*.cs     diff=csharp
+*.bat        text eol=crlf
+#*.bat        text eol=lf
+*.gradle     text diff=java
+mcmod.info   text
+*.java       text
+*.json       text
+*.lang       text
+*.mcmeta     text
+*.md         text
+*.properties text
+gradlew      text eol=lf
+*.sh         text eol=lf
+*.txt        text
+*.xml        text
+*.yml        text
+*.yaml       text
+*.toml       text
 
-# Standard to msysgit
-*.doc	 diff=astextplain
-*.DOC	 diff=astextplain
-*.docx diff=astextplain
-*.DOCX diff=astextplain
-*.dot  diff=astextplain
-*.DOT  diff=astextplain
-*.pdf  diff=astextplain
-*.PDF	 diff=astextplain
-*.rtf	 diff=astextplain
-*.RTF	 diff=astextplain
+# Documents
+*.pdf  binary
+
+# Graphics
+*.eps  binary
+*.gif  binary
+*.ico  binary
+*.jpg  binary
+*.jpeg binary
+*.png  binary
+*.psd  binary
+# SVG treated as an asset (binary) by default. If you want to treat it as text,
+# comment-out the following line and uncomment the line after.
+*.svg  binary
+#*.svg text
+*.tif  binary
+*.tiff binary
+*.webp binary
+*.xcf  binary
+
+# Other
+*.exe binary
+*.jar binary
+
+###############################
+# Git Large File System (LFS) #
+###############################
+
+# Archives
+#*.7z   filter=lfs diff=lfs merge=lfs -text
+#*.br   filter=lfs diff=lfs merge=lfs -text
+#*.bz2  filter=lfs diff=lfs merge=lfs -text
+#*.gz   filter=lfs diff=lfs merge=lfs -text
+#*.tar  filter=lfs diff=lfs merge=lfs -text
+#*.zip  filter=lfs diff=lfs merge=lfs -text
+
+# Documents
+#*.pdf  filter=lfs diff=lfs merge=lfs -text
+
+# Graphics
+#*.eps  filter=lfs diff=lfs merge=lfs -text
+#*.gif  filter=lfs diff=lfs merge=lfs -text
+#*.ico  filter=lfs diff=lfs merge=lfs -text
+#*.jpg  filter=lfs diff=lfs merge=lfs -text
+#*.jpeg filter=lfs diff=lfs merge=lfs -text
+#*.png  filter=lfs diff=lfs merge=lfs -text
+#*.psd  filter=lfs diff=lfs merge=lfs -text
+#*.tif  filter=lfs diff=lfs merge=lfs -text
+#*.tiff filter=lfs diff=lfs merge=lfs -text
+#*.webp filter=lfs diff=lfs merge=lfs -text
+#*.xcf  filter=lfs diff=lfs merge=lfs -text
+
+# Other
+#*.exe  filter=lfs diff=lfs merge=lfs -text
+#*.jar  filter=lfs diff=lfs merge=lfs -text
