+++ .gitignore
@@ -1,20 +1,35 @@
-*.class
-*.iml
+# eclipse
+bin
+*.launch
+.settings
+.settings/*
+.metadata
+.classpath
+.project
+
+# idea
+out
 *.ipr
 *.iws
-*.classpath
-*.project
-*.prefs
-*.launch
-/output
+*.iml
+.idea
 
-#ForgeGradle
-/build
-/.gradle
-/eclipse/assets
-/libs
-/bin/
-/run/
-/out/
+# gradle
+build
+.gradle
 
-\.idea/
+# other
+eclipse
+run
+classes
+*.hprof
+
+/README.txt
+/forge-*-changelog.txt
+
+secret.json
+libs/*
+!/libs/README.txt
+
+.DS_Store
+.okhttpcache/*
