+++ changelog.mustache
@@ -1,0 +1,10 @@
+{{#tags}}
+ {{#issues}}
+  {{#commits}}
+**{{{messageTitle}}}**
+{{#messageBodyItems}}
+ * {{.}}
+{{/messageBodyItems}}
+  {{/commits}}
+ {{/issues}}
+{{/tags}}
