+++ gradle.properties
@@ -1,11 +1,101 @@
+mod_version=2.2.0
+org.gradle.daemon = false
 
-# Sets default memory used for gradle commands. Can be overridden by user or command line properties.
-# This is required to provide enough memory for the Minecraft decompilation process.
-org.gradle.jvmargs=-Xmx3G
-org.gradle.daemon=false
-mc_version=1.16.4
-forge_version=35.1.2
-mapping=20201028
-jei_version=7.0.0.6
-mod_version=2.1
-modid=ropebridge
+group = com.mcmoddev
+archivesBaseName = RopeBridge
+name = Rope Bridge
+description = A mod that adds rope bridges to Minecraft. Craft the grappling gun and shoot it over a gap large or small to build a rope bridge across it.
+
+#Minumum Minecraft Version Mod will be built with
+mc_version=1.16.5
+
+#ForgeGradle version to use (Required)
+fg_version = 4.1.+
+#version of Minecraft Forge is built against (Required)
+forge_mc_version = 1.16.5
+#Forge version to use (Required)
+forge_version = 36.1.23
+#Mappings channel: snapshot or stable
+mcp_mappings_channel = snapshot
+#Mappings version to use (Required)
+mcp_mappings_version = 20210309-1.16.5
+
+#Cursforge Project ID (Required)
+curseforge_project_id = 238050
+#Type of Release, can be one of 'alpha', 'beta' or 'release' (Defaults to 'alpha' if not set)
+curseforge_release_type = release
+#When using a ChangeLog file specifies filename
+curseforge_changelog_filename = CHANGELOG.txt
+#Changelog type, can be one of 'text', 'html' or 'markdown' (Defaults to 'text' if not set)
+curseforge_changelog_type = text
+#Version(s) of Minecraft this mod will work on (comma separated list)
+curseforge_versions = 1.16.5
+#Whether to use a custom display name on artifacts
+curseforge_use_custom_display_name = true
+#List of required dependencies (comma separated list)
+#curseforge_requirements =
+#List of optional dependencies (comma separated list)
+#curseforge_optionals =
+#List of embedded dependencies (comma separated list)
+#curseforge_embeddeds =
+#List of compatible tools (comma separated list)
+#curseforge_tools =
+#List of incompatible dependencies (comma separated list)
+#curseforge_incompatibles =
+#List of included dependencies (comma separated list)
+#curseforge_includes =
+
+#String reference to Core Plugin this mod contains, if any
+#core_plugin =
+#Whether or not to use Access Transformers from depended mods
+dep_has_ats = false
+#File name of this mod's Access Transformers (If any)
+#mod_at_file = _at.cfg
+
+#Whether the artifacts should be signed
+do_sign_jar = true
+
+#Create a source jar?
+create_source_jar = true
+#Create an API jar?
+create_api_jar = false
+#Create a deobf jar?
+create_deobf_jar = false
+#Create a javadoc jar?
+create_javadoc_jar = true
+
+#The following four options are optional and do not affect your build in any way
+#They are only for convenience when using the 'runClient' and 'runServer' tasks
+#Whether to Disable server gui
+#mc_server_nogui = true
+#Whether to Skip the screen to confirm that you want to load a world with missing registry entries
+#forge_do_not_backup = true
+#Whether to skip having to confirm on server
+#forge_query_result_confirm = true
+#Whether to Skip jansi warnings in the log
+#log4j_skip_jansi = true
+
+# Optional convenience setter for game resolution one of '480p', '576p', '720p', '1080p' or 'custom'
+#mc_resolution = 1080p
+#the following two will only be used when mc_resolution = custom
+#mc_custom_resolution_width = 1920
+#mc_custom_resolution_height = 1080
+#mc_fullscreen = true
+
+#Additional arguments to pass to minecraft.clientJvmArgs
+#client_jvm_args =
+#Additional arguments to pass to minecraft.serverJvmArgs
+#client_game_args =
+#Additional arguments to pass to minecraft.clientRunArgs
+#server_jvm_args =
+#Additional arguments to pass to minecraft.serverRunArgs
+#server_game_args =
+
+##The following three options should *NEVER* be set in this file, it's a security risk, Be safe, keep your account information private.
+##These properties here are for reference only and explanation of what they do
+##Your Minecraft account username (Consistent player name for singleplayer worlds)
+##mc_username =
+##Your Minecraft account password (Allows online play while present with associated username)
+##mc_password =
+##Your Minecraft account uuid (Allows using your skin while testing, This can be with or without hyphens)
+##mc_uuid =
