--- setup.bat
@@ -1,2 +1,0 @@
-call gradlew setupDecompWorkspace --refresh-dependencies eclipse
-pause
