+++ src/main/resources/META-INF/mods.toml
@@ -6,10 +6,10 @@
 # The savedName of the mod loader type to load - for regular FML @Mod mods it should be javafml
 modLoader="javafml" #mandatory
 # A version range to match for said mod loader - for regular FML @Mod it will be the forge version
-loaderVersion="[28,)" #mandatory (24 is current forge version)
+loaderVersion="[36,)" #mandatory (24 is current forge version)
 # A URL to refer people to when problems occur with this mod
-#issueTrackerURL="https://github.com/" #optional
-license="All Rights Reserved"
+#issueTrackerURL="https://github.com/MinecraftModDevelopmentMods/Rope-Bridge" #optional
+license="GNU LESSER GENERAL PUBLIC LICENSE Version 2.1"
 # A list of mods - how many allowed here is determined by the individual mod loader
 [[mods]] #mandatory
 # The modid of the mod
@@ -19,18 +19,16 @@ version="${file.jarVersion}" #mandatory
  # A display savedName for the mod
 displayName="Rope Bridge" #mandatory
 # A URL to query for updates for this mod. See the JSON update specification <here>
-#updateJSONURL="http://myurl.me/" #optional
+updateJSONURL="https://raw.githubusercontent.com/MinecraftModDevelopmentMods/Rope-Bridge/dev/updates.json" #optional
 # A URL for the "homepage" for this mod, displayed in the mod UI
-displayURL="https://minecraft.curseforge.com/projects/rope-bridge-mod/" #optional
+displayURL="https://www.curseforge.com/minecraft/mc-mods/rope-bridge" #optional
 # A file savedName (in the root of the mod JAR) containing a logo for display
 logoFile="ropebridge.logo.png" #optional
 # A text field displayed in the mod UI
 credits="Originally created by czechmate777. Art from wiresgal and TrisAstral" #optional
 # A text field displayed in the mod UI
-authors="MrTrollNugNug, Lewis_McReu, Darkhax" #optional
+authors="MrTrollNugNug, Lewis_McReu, Darkhax, noobanidus, Jriwanek, Jaredlll08, yrsegal, tfarcenim, ProxyNeko" #optional
 # The description text for the mod (multi line!) (#mandatory)
 description='''
-
 A mod that adds rope bridges to Minecraft. Craft the grappling gun and shoot it over a gap large or small to build a rope bridge across it.
-
 '''
