+++ src/main/resources/pack.mcmeta
@@ -1,6 +1,6 @@
 {
     "pack": {
-        "description": "examplemod resources",
+        "description": "rope bridge mod's resources",
         "pack_format": 4,
         "_comment": "A pack_format of 4 requires json lang files. Note: we require v4 pack meta for all mods."
     }
