+++ updates.json
@@ -1,0 +1,74 @@
+{
+  "homepage": "https://www.curseforge.com/minecraft/mc-mods/rope-bridge",
+  "promos": {
+    "1.16.5-latest": "2.2.0",
+    "1.16.1-latest": "2.0",
+    "1.15.1-latest": "1.3",
+    "1.12.2-latest": "2.0.7",
+    "1.12.1-latest": "2.0.6",
+    "1.11.2-latest": "2.0.6",
+    "1.10.2-latest": "2.0.5",
+    "1.8.9-latest": "v1.3",
+    "1.8.8-latest": "v1.3",
+    "1.8-latest": "v1.3",
+    "1.7.10-latest": "v1.3",
+    "1.16.5-recommended": "2.2.0",
+    "1.16.1-recommended": "2.0",
+    "1.15.1-recommended": "1.3",
+    "1.12.2-recommended": "2.0.7",
+    "1.12.1-recommended": "2.0.6",
+    "1.11.2-recommended": "2.0.6",
+    "1.10.2-recommended": "2.0.5",
+    "1.8.9-recommended": "v1.3",
+    "1.8.8-recommended": "v1.2",
+    "1.8-recommended": "v1.3",
+    "1.7.10-recommended": "v1.3"
+  },
+  "1.16.5": {
+    "2.2.0": "Refactored code, fixed items not being consumed on bridge creation, updated tooling for development.",
+    "2.1": "Initial port to 1.16.4/5"
+  },
+  "1.16.1": {
+    "2.0": "Initial port to 1.16.1"
+  },
+  "1.15.1": {
+    "1.3": "Fix #32, calling the wrong particle method on the server",
+    "1.2": "Added the Russian translation file.",
+    "1.1": "Initial port to 1.15.1"
+  },
+  "1.12.2": {
+    "2.0.7": "Initial port to 1.12.2"
+  },
+  "1.12.1": {
+    "2.0.6": "Initial port to 1.12.1"
+  },
+  "1.11.2": {
+    "2.0.6": "Changelogs not kept.",
+    "2.0.5": "Changelogs not kept.",
+    "1.4.8": "Initial port to 1.11.2"
+  },
+  "1.10.2": {
+    "2.0.5": "Changelogs not kept.",
+    "2.0.1": "Changelogs not kept.",
+    "2.0.0": "Changelogs not kept.",
+    "1.4.8": "Changelogs not kept.",
+    "1.4.6": "Changelogs not kept.",
+    "1.4.5": "Changelogs not kept.",
+    "1.4.0": "Changelogs not kept.",
+    "1.3": "Initial port to 1.10.2"
+  },
+  "1.8.9": {
+    "v1.3": "Initial port to 1.8.9"
+  },
+  "1.8.8": {
+    "v1.2": "Initial port to 1.8.8"
+  },
+  "1.8": {
+    "v1.3": "Changelogs not kept.",
+    "v1.2": "Initial port to 1.8."
+  },
+  "1.7.10": {
+    "v1.3": "Secondary release.",
+    "v1.2": "Initial release."
+  }
+}
